change location and transfer some functionalities from map-screen
